feat(fe-quality): abstract duplicated form title input into its own reusable component
